requirements.txt
Flask==2.2.5
firebase-admin==6.3.0
python-dotenv==1.0.0
mysql-connector-python==9.0.0
Flask-Caching==2.0.2
Flask-Login==0.6.2
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
pandas==1.4.3 
numpy==1.22.4
pytz==2023.3
